requirements.txt
app_store==1.0.1
atlassian==0.0.0
beautifulsoup4==4.12.3
dateparser==1.2.0
elasticsearch==8.13.0
Flask==3.0.3
google_api_python_client==2.126.0
google_play_scraper==1.2.6
GoogleNews==1.6.14
Jinja2==3.1.3
mmh3==4.1.0
nltk==3.8.1
pandas==1.4.0
praw==7.7.1
presidio_analyzer==2.2.354
presidio_anonymizer==2.2.354
protobuf==3.20.3
pydantic==2.7.0
pydantic_settings==2.2.1
pyfacebook==0.3
python_dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.1
reddit_rss_reader==1.3.2
Requests==2.31.0
searchtweets==1.7.6
setuptools==57.4.0
slack_sdk==3.27.1
spacy==3.7.4
SQLAlchemy==2.0.29
torch==2.2.2
trafilatura==1.8.1
transformers==4.39.3
urllib3==2.2.1
vaderSentiment==3.3.2
